Renaming component folder
